requirements.txt
# ============================================================
# AI Strike Team — Python Dependencies
# ============================================================

# --- LLM ---
openai>=1.30.0
anthropic>=0.25.0

# --- Orchestration / Agents ---
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0

# --- Validation ---
pydantic>=2.7.0
pydantic-settings>=2.2.0

# --- File Ingestion ---
pypdf>=4.2.0           # PDF parsing
python-docx>=1.1.0     # DOCX parsing
openpyxl>=3.1.2        # Excel parsing
python-multipart>=0.0.9

# --- Web / API ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# --- Database ---
sqlalchemy>=2.0.0
asyncpg>=0.29.0        # async Postgres driver
alembic>=1.13.0        # DB migrations

# --- Email ---
imapclient>=3.0.1

# --- Logging ---
structlog>=24.1.0

# --- Utilities ---
python-dotenv>=1.0.0
tenacity>=8.3.0        # retry logic
rich>=13.7.0           # pretty CLI output
typer>=0.12.0          # CLI runner
pyyaml>=6.0.1
pandas>=2.2.0          # data analysis
jinja2>=3.1.0          # report templates
slack-sdk>=3.27.0      # Slack notifications

# --- Dev / Testing ---
pytest>=8.2.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
